Redid input and made conversion from random values for each keys to a new array which is contiguous
The code now instead of having an array which stores each offset between the different keysets, now has a new array which stores the difference between the values and also the distance between each block of keys.

This allows for the mask to be converted back to the ASCII values but also allows for the keys to be stored in a mask of two longs which in total is 128 bytes.
This was sort of achieved in the previous commit however it relied on a value in the camera struct which stored which offset you were in.

However the new system allows for the section to be checked without this variable which means the camera won't get stuck if you suddenly press Space after moving with the WASD keys